Refactor `Result` from class to interface
