+++ instancio-tests/instancio-core-tests/src/test/java/org/instancio/internal/InternalResultTest.java
@@ -1,0 +1,32 @@
+/*
+ * Copyright 2022-2024 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.instancio.internal;
+
+import org.junit.jupiter.api.Test;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+class InternalResultTest {
+
+    @Test
+    void verifyToString() {
+        assertThat(new InternalResult<>("foo", 123L))
+                .hasToString("Result[seed=123, object=foo]");
+
+        assertThat(new InternalResult<>(null, 123L))
+                .hasToString("Result[seed=123, object=null]");
+    }
+}
